Switched PKs from long (bigint) to int.  Allowed for null values in WaiverID, ServiceTypeID, and LocationID in Project Table.

+++ v1-to-v2-data-conversion/v1-to-v2-data-conversion/Data/v2/Employee.cs
@@ -10,10 +10,10 @@ namespace Conversion.Data.v2
             WorkOpportunityForEmployee = new HashSet<WorkOpportunityForEmployee>();
         }
 
-        public long EmployeeId { get; set; }
+        public int EmployeeId { get; set; }
         public string EmployeeFirstName { get; set; }
         public string EmployeeLastName { get; set; }
-        public long LocationId { get; set; }
+        public int LocationId { get; set; }
         public string EmployeeEmail { get; set; }
         public string EmployeePhone { get; set; }
         public string EmployeeCellPhone { get; set; }
