+++ v1-to-v2-data-conversion/v1-to-v2-data-conversion/Data/v2/Location.cs
@@ -11,7 +11,7 @@ namespace Conversion.Data.v2
             Project = new HashSet<Project>();
         }
 
-        public long LocationId { get; set; }
+        public int LocationId { get; set; }
         public string LocationName { get; set; }
         public string LocationCountry { get; set; }
